requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
aiosignal==1.2.0
anytree==2.8.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
asynctest==0.13.0
atari-py==0.2.9
attrs==21.2.0
blessed==1.20.0
blessings==1.7
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.1
cloudpickle==1.6.0
cmake==3.18.4.post1
colorama==0.4.4
colorful==0.5.4
commentjson==0.9.0
contextlib2==21.6.0
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dm-sonnet==1.35
dm-tree==0.1.6
drawnow==0.72.5
filelock==3.0.12
flatbuffers==1.12
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-api-core==1.28.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.4.1
graph-nets==1.0.4
grpcio==1.34.1
gym==0.18.0
h5py==3.1.0
hiredis==2.0.0
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.0.1
iniconfig==1.1.1
jsonschema==3.2.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lark-parser==0.7.8
llvmlite==0.36.0
lz4==3.1.3
Markdown==3.3.4
matplotlib==3.3.4
mfem==4.4.0.0
msgpack==1.0.2
multidict==5.1.0
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
nvidia-ml-py==12.535.77
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.5.2.52
opencv-python-headless==4.3.0.36
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
Pillow==7.2.0
pluggy==1.0.0
prometheus-client==0.10.1
protobuf==3.17.1
psutil==5.8.0
py==1.11.0
py-spy==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyglet==1.5.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==7.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
ray==1.5.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.2
semantic-version==2.10.0
six==1.15.0
smmap==5.0.0
svgwrite==1.4.1
tabulate==0.8.9
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
tomli==1.2.3
torch==1.8.1
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.6
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
